refactor: return R1CS/layer compression and commit logging to the spine
CompressedR1CS, CompressedLayers, and log_commit_input were field-agnostic
(postcard byte compression of R1CS<F>/witness layers; nonzero-density
logging over any Field) but had landed in provekit-backend-bn254 during the
glue move simply because they shared a file with the acir witness solver.

Move them to provekit-common: CompressedR1CS::compress/decompress are now
generic over <F: Field>; CompressedLayers and log_commit_input<F> move as-is.
Backend keeps only the genuinely acir-welded solve_witness_vec and the
test-only R1CSSolver. bn254 output unchanged (72/72 e2e green).
